Started on Rust Unbounded Channel implementation.
Some items that are of note are:
  - It uses the crossbeam crate's epoch-based memory reclamation system to handle
    freeing memory.
  - The writeup mentions that the way in which the Go implementation manages
    handles on the channel is prone to space leaks. The Rust implementation uses
    a different design for handles that allows one to safely update head and
    tail pointers on the global struct.
  - Because Rust programs are less likely to have tens of thousands+ of
    (potentially relatively short-lived) threads, the allocation cache is
    implemented as a simple thread-local Option, rather than a global list.

There is a binary that runs a standard enqueue/dequeue microbenchmark for this
implementation (fchan) and crossbeam's implementation of a blocking
Michael-Scott Queue. The bench.rs is entirely safe Rust, so any segfaults/funny
business there should point to bugs in the underlying data-structure
implementations.

Unlike Go --- where the atomic operations all provide sequential consistency ---
Rust's atomic operations follow a subset of the LLVM memory orderings. I believe
that relaxed ordering is sufficient most of the time, but that there indeed
points where barriers are necessary. I am not confident that I have gotten all
of these places right; I am probably too loose with ordering in some places and
too strict in others.